criando componente para cores de agendaento
